refactor: part 1
