+++ storage-proofs/src/circuit/mod.rs
@@ -1,21 +1,4 @@
-pub mod constraint;
-mod insertion;
-
-pub mod create_label;
 pub mod drgporep;
-pub mod election_post;
-pub mod encode;
 pub mod multi_proof;
-pub mod multipack;
-pub mod pedersen;
 pub mod por;
-pub mod rational_post;
 pub mod stacked;
-pub mod uint64;
-pub mod variables;
-pub mod xor;
-
-// FIXME: Can we make a config like for test?
-pub mod bench;
-pub mod metric;
-pub mod test;
