+++ storage-proofs/src/gadgets/variables.rs
