+++ storage-proofs/src/lib.rs
@@ -12,8 +12,6 @@ pub mod compound_proof;
 pub mod crypto;
 pub mod drgporep;
 pub mod drgraph;
-pub mod election_post;
-mod encode;
 pub mod error;
 pub mod fr32;
 pub mod hasher;
@@ -24,12 +22,19 @@ pub mod parameter_cache;
 pub mod partitions;
 pub mod pieces;
 pub mod porep;
+pub mod post;
 pub mod proof;
-pub mod rational_post;
 pub mod sector;
 pub mod settings;
 pub mod stacked;
 pub mod util;
+
+pub mod gadgets;
+
+mod data;
+mod encode;
+
+pub use self::data::Data;
 
 pub(crate) const TEST_SEED: [u8; 16] = [
     0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc, 0xe5,
