+++ storage-proofs/src/post/election/vanilla.rs
