+++ storage-proofs/src/post/rational/mod.rs
@@ -1,0 +1,7 @@
+mod circuit;
+mod compound;
+mod vanilla;
+
+pub use self::circuit::*;
+pub use self::compound::*;
+pub use self::vanilla::*;
