+++ storage-proofs/src/post/rational/vanilla.rs
